+++ packages/flutter/lib/src/physics/clamped_simulation.dart
@@ -1,6 +1,8 @@
-// Copyright 2016 The Chromium Authors. All rights reserved.
+// Copyright 2014 The Flutter Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
+
+// @dart = 2.8
 
 import 'simulation.dart';
 
@@ -48,10 +50,10 @@ class ClampedSimulation extends Simulation {
   final double dxMax;
 
   @override
-  double x(double time) => simulation.x(time).clamp(xMin, xMax);
+  double x(double time) => simulation.x(time).clamp(xMin, xMax) as double;
 
   @override
-  double dx(double time) => simulation.dx(time).clamp(dxMin, dxMax);
+  double dx(double time) => simulation.dx(time).clamp(dxMin, dxMax) as double;
 
   @override
   bool isDone(double time) => simulation.isDone(time);
